// seller/src/component/ecommerce/product/create-edit/form-utils.ts
import { CreateProductInput } from '@/utils/validators/create-product.schema';

export const customFields = [
  {
    label: '',
    value: '',
  },
];
export const locationShipping = [
  {
    name: '',
    shippingCharge: '',
  },
];
export const productVariants = [
  {
    label: '',
    value: '',
  },
];

export function defaultValues(product?: CreateProductInput) {
  return {
    name: product?.name ?? '',
    price: product?.price ?? undefined,
    mrp: product?.mrp ?? undefined,
    brand: product?.brand ?? '',
    category: product?.category ?? '',
    desc: product?.desc ?? '',
    images: product?.images ?? undefined,
    stock: product?.stock ?? '',
    tags: product?.tags ?? [],
    colors: product?.colors || [],
    sizes: product?.sizes || [],
    instaId: product?.instaId ?? '',
    height: product?.height ?? '',
    weight: product?.weight ?? '',
    length: product?.length ?? '',
    breadth: product?.breadth ?? '',
  };
}

export const productData = {
  title: 'Apple',
  description: 'Fresh Express Iceberg Garden Salad Blend',
  sku: 'SKU-28935',
  type: 'Digital Product',
  categories: 'Grocery',
  price: 10,
  costPrice: 20,
  retailPrice: 15,
  salePrice: 25,
  productImages: undefined,
  inventoryTracking: 'no',
  currentStock: '150',
  lowStock: '20',
  productAvailability: 'online',
  tradeNumber: '12345',
  manufacturerNumber: '154',
  brand: 'Foska',
  upcEan: 'Ean',
  customFields: [
    {
      label: 'Color',
      value: 'Red',
    },
  ],
  freeShipping: false,
  shippingPrice: 45,
  locationBasedShipping: true,
  locationShipping: [
    {
      name: 'USA',
      shippingCharge: '150',
    },
  ],
  pageTitle: 'apple',
  metaDescription: 'apple',
  metaKeywords: 'grocery, foods',
  productUrl: 'http://localhost:3000/',
  isPurchaseSpecifyDate: true,
  isLimitDate: true,
  dateFieldName: 'Date Field',
  productVariants: [
    {
      name: 'Jhon',
      value: '150',
    },
  ],
  tags: ['iPhone', 'mobile'],
};

export const menuItems = [
  {
    label: 'Summary',
    value: 'summary',
  },
  {
    label: 'Images & Gallery',
    value: 'images_gallery',
  },
  {
    label: 'Pricing & Inventory',
    value: 'pricing_inventory',
  },
  {
    label: 'Product Identifiers & Custom Fields',
    value: 'product_identifiers',
  },
  {
    label: 'Shipping & Availability',
    value: 'shipping_availability',
  },
  {
    label: 'SEO',
    value: 'seo',
  },
  {
    label: 'Variant Options',
    value: 'variant_options',
  },
];

// Category option
export const categoryOption = [
  {
    value: 'fruits',
    name: 'Fruits',
  },
  {
    value: 'grocery',
    name: 'Grocery',
  },
  {
    value: 'meat',
    name: 'Meat',
  },
  {
    value: 'cat food',
    name: 'Cat Food',
  },
];

// Type option
export const typeOption = [
  {
    value: 'digital product',
    name: 'Digital Product',
  },
  {
    value: 'physical product',
    name: 'Physical Product',
  },
];

// Variant option
export const variantOption = [
  {
    value: 'single',
    name: 'Single',
  },
  {
    value: 'multiple',
    name: 'Multiple',
  },
];
